GUI: Updates to home screen, added initial Choose Simulations screen
New home screen now checks if the map is indeed in the Database, and if clicking Learning mode or Compare it goes to a Verify Map screen and then to the next matching screen. Code updated in DrawMapController to implement this.
Also fixed important typo bug in CitySimulation, evaluator wasn't written correctly in the constructor so it was effectively ignored.

issue #141
issue #142