GUI: Prepare Config updated (save new map), BTWDatabase update (saveMap, parseMap)
Now on prepare config screen we can save a new generated map. A white box will appear in the middle until the saving is over. the user must insert a map name.
Changes on the BTWDatabase: parseMap used to also save the map in a separate thread. I couldn't not save a map I didn't accept. Now saving the map (including createStatisticsTable) was moved to saveMap. saveMap needs to be called for a map to be saved in the database. I also added a second saveMap which doesn't get parameters. It remembers the geoJson map from calling parseMap.

issue #147
issue #84